PDA_upload_s3/requirements.txt
boto3>=1.34
botocore>=1.34
python-dotenv>=1.0
tqdm>=4.66
